test_driver: use domain-private data to store managed image
The managedSave APIs according to the documentation are supposed to
operate on a disk file. However, this might not be appropriate in the
case of the test driver since:

* It's better if the test driver keeps all its state in memory only and
doesn't affect the host in any way.

* The test driver, apart from "emulating" the domains, it additionally
"emulates" a fake physical host. Every time we start a new test
connection that sort of means that a new physical host is created as
well. And this fake host isn't necessarily the same.

What we can do instead is operating on the already existing domain
definitions. So along as a connection remains open, a domain can
preserve the managed state between different shutdown / create calls.
When the test connection closes this means that the fake host is
destroyed as well, hence no other state is preserved after that.

This way we also make sure that we don't touch the real host's
filesystem.

Signed-off-by: Ilias Stamatis <stamatis.iliass@gmail.com>